Add LoadingSystem, add Loadable and Loaded components to entities

+++ src/systems/kill_enemy.rs
@@ -9,6 +9,8 @@ impl<'a> System<'a> for KillEnemySystem {
         ReadStorage<'a, Player>,
         ReadStorage<'a, Enemy>,
         ReadStorage<'a, Collision>,
+        ReadStorage<'a, Loadable>,
+        ReadStorage<'a, Loaded>,
         WriteStorage<'a, Velocity>,
         WriteStorage<'a, Spike>,
     );
@@ -20,6 +22,8 @@ impl<'a> System<'a> for KillEnemySystem {
             players,
             enemies,
             collisions,
+            loadables,
+            loadeds,
             mut velocities,
             mut spikes,
         ): Self::SystemData,
@@ -28,17 +32,34 @@ impl<'a> System<'a> for KillEnemySystem {
             (&players, &collisions, &mut velocities).join().next()
         {
             if player_velocity.y < 0.0 {
-                for (enemy_entity, _, enemy_spike_opt) in
-                    (&entities, &enemies, (&mut spikes).maybe()).join()
+                for (
+                    enemy_entity,
+                    _,
+                    enemy_spike_opt,
+                    loadable_opt,
+                    loaded_opt,
+                ) in (
+                    &entities,
+                    &enemies,
+                    (&mut spikes).maybe(),
+                    loadables.maybe(),
+                    loadeds.maybe(),
+                )
+                    .join()
                 {
-                    if player_collision.in_collision_with(enemy_entity.id()) {
-                        // Kill enemy and make the player bounce off their head
-                        player_velocity.y = player.settings.bounce_strength;
-                        entities
-                            .delete(enemy_entity)
-                            .expect("Tried to kill enemy");
-                        if let Some(enemy_spike) = enemy_spike_opt {
-                            enemy_spike.enabled = false;
+                    if let (None, None) | (Some(_), Some(_)) =
+                        (loadable_opt, loaded_opt)
+                    {
+                        if player_collision.in_collision_with(enemy_entity.id())
+                        {
+                            // Kill enemy and make the player bounce off their head
+                            player_velocity.y = player.settings.bounce_strength;
+                            entities
+                                .delete(enemy_entity)
+                                .expect("Tried to kill enemy");
+                            if let Some(enemy_spike) = enemy_spike_opt {
+                                enemy_spike.enabled = false;
+                            }
                         }
                     }
                 }
